added pagination for clients

+++ src/main/java/by/kynca/estateWeb/repository/RealEstateRepo.java
@@ -11,6 +11,6 @@ import java.util.List;
 public interface RealEstateRepo extends JpaRepository<RealEstate, Long> {
     Page<RealEstate> findRealEstatesByRealEstateType(RealEstateType realEstateType, Pageable pageable);
 
-    long countByRealEstateType(RealEstateType realEstateType);
+    int countByRealEstateType(RealEstateType realEstateType);
     List<RealEstate> findRealEstatesByClientClientId(Long clientId);
 }
